refactor: 컴포넌트들을 components 폴더로, 공통 컴포넌트는 components/common으로 이동 #7
