+++ src/components/common/Input.jsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const Input = ({ type, id, name, label, value, onChange, step }) => {
     return (
         <label htmlFor={id}>
